pyproject.toml
[tool.poetry]
name = "stiltctl"
version = "0.1.0"
description = "Run distributed STILT workflows."
authors = ["Ben Fasoli <benfasoli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.8.2"
typer = "^0.3.2"
netCDF4 = "^1.5.7"
loguru = "^0.5.3"
python-dotenv = "^0.19.0"
cloudstorage = {extras = ["google"], version = "^0.11.0"}
matplotlib = "^3.4.3"
python-magic = "^0.4.24"
filelock = "^3.3.0"
itsdangerous = "^2.0.1"
xattr = "^0.9.7"
PyYAML = "^5.4.1"
SQLAlchemy = "^1.4.25"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
mypy = "^0.910"
pytest = "^5.2"
pytest-cov = "^2.12.1"
pytest-profiling = "^1.7.0"
isort = "^5.9.1"
autoflake = "^1.4"
cloudstorage = {extras = ["local"], version = "^0.11.0"}
types-PyYAML = "^5.4.11"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.25"}
migra = "^3.0.1621480950"

[tool.poetry.scripts]
stiltctl = "stiltctl.cli:app"

[tool.isort]
profile = "black" 
src_paths = ["stiltctl", "tests"]

[tool.mypy]
plugins = ["sqlalchemy.ext.mypy.plugin"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
